refactor and combine search tests
